[IBDR-145] Update response creation

+++ brapi/src/main/java/org/planth_pheno_analytics/brapi/api/response/metadata/Pagination.java
@@ -17,7 +17,7 @@ public class Pagination {
         this.totalPages = totalPages;
     }
 
-    public static Pagination of(int currentPage, int pageSize, int totalCount, int totalPages) {
+    public static Pagination of(int currentPage, int pageSize, long totalCount, int totalPages) {
         return new Pagination(currentPage, pageSize, totalCount, totalPages);
     }
